feat : 예외 로직 수정

+++ src/main/java/com/sparta/fmdelivery/domain/review/service/ReviewService.java
@@ -1,5 +1,6 @@
 package com.sparta.fmdelivery.domain.review.service;
 
+import com.sparta.fmdelivery.apipayload.status.ErrorStatus;
 import com.sparta.fmdelivery.domain.common.dto.AuthUser;
 import com.sparta.fmdelivery.domain.review.dto.ReviewRequest;
 import com.sparta.fmdelivery.domain.review.dto.ReviewResponse;
@@ -11,6 +12,7 @@ import com.sparta.fmdelivery.domain.shop.entitiy.Shop;
 import com.sparta.fmdelivery.domain.shop.repository.ShopRepository;
 import com.sparta.fmdelivery.domain.user.entity.User;
 import com.sparta.fmdelivery.domain.user.repository.UserRepository;
+import com.sparta.fmdelivery.exception.ApiException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -35,16 +37,12 @@ public class ReviewService {
      */
     @Transactional
     public ReviewResponse createReview(AuthUser authUser, ReviewRequest request) {
-        Order order = orderRepository.findById(request.getOrderId())
-                .orElseThrow(() -> new IllegalArgumentException("주문을 찾을 수 없습니다."));
-        Shop shop = shopRepository.findById(request.getShopId())
-                .orElseThrow(() -> new IllegalArgumentException("가게를 찾을 수 없습니다."));
-        User user = userRepository.findById(authUser.getId())
-                .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
-        Review review = new Review(request, order, shop, user);
+        Order order = getOrderById(request.getOrderId());
+        Shop shop = getShopById(request.getShopId());
+        User user = getUserById(authUser.getId());
 
-        Review savedReview = reviewRepository.save(review);
-        return ReviewResponse.fromEntity(savedReview);
+        Review review = new Review(request, order, shop, user);
+        return ReviewResponse.fromEntity(reviewRepository.save(review));
     }
 
     /**
@@ -52,24 +50,20 @@ public class ReviewService {
      * @param reviewId
      * @return
      */
-    @Transactional
+    @Transactional(readOnly = true)
     public ReviewResponse getReview(Long reviewId) {
-        Review review = reviewRepository.findById(reviewId)
-                .orElseThrow(() -> new IllegalArgumentException("리뷰를 찾을 수 없습니다."));
-
+        Review review = getReviewById(reviewId);
         return ReviewResponse.fromEntity(review);
     }
-
 
     /**
      * 해당 가게의 전체 리뷰 조회
      * @param shopId
      * @return
      */
-    @Transactional
+    @Transactional(readOnly = true)
     public List<ReviewResponse> getReviews(Long shopId) {
         List<Review> reviews = reviewRepository.findAllByShopId(shopId);
-
         return reviews.stream()
                 .map(ReviewResponse::fromEntity)
                 .collect(Collectors.toList());
@@ -81,11 +75,10 @@ public class ReviewService {
      * @param request
      * @return
      */
+    @Transactional
     public ReviewResponse updateReview(AuthUser authUser, Long reviewId, ReviewRequest request) {
-        isValidUser(reviewId, authUser.getId());
-
-        Review review = reviewRepository.findById(reviewId)
-                .orElseThrow(() -> new IllegalArgumentException("리뷰를 찾을 수 없습니다."));
+        Review review = getReviewById(reviewId);
+        validateUserPermission(review.getUser().getId(), authUser.getId());
 
         review.updateReview(request.getRating(), request.getContent());
         return ReviewResponse.fromEntity(review);
@@ -98,22 +91,60 @@ public class ReviewService {
      */
     @Transactional
     public void deleteReview(AuthUser authUser, Long reviewId) {
-        isValidUser(reviewId, authUser.getId());
+        Review review = getReviewById(reviewId);
+        validateUserPermission(review.getUser().getId(), authUser.getId());
 
-        Review review = reviewRepository.findById(reviewId)
-                        .orElseThrow(() -> new IllegalArgumentException("리뷰를 찾을 수 없습니다."));
         reviewRepository.delete(review);
     }
 
     /**
-     * 리뷰의 사용자 권한 확인
+     * 주문 조회
+     * @param orderId
+     * @return Order
+     */
+    private Order getOrderById(Long orderId) {
+        return orderRepository.findById(orderId)
+                .orElseThrow(() -> new ApiException(ErrorStatus._NOT_FOUND_ORDER));
+    }
+
+    /**
+     * 가게 조회
+     * @param shopId
+     * @return Shop
+     */
+    private Shop getShopById(Long shopId) {
+        return shopRepository.findById(shopId)
+                .orElseThrow(() -> new ApiException(ErrorStatus._NOT_FOUND_SHOP));
+    }
+
+    /**
+     * 사용자 조회
+     * @param userId
+     * @return User
+     */
+    private User getUserById(Long userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new ApiException(ErrorStatus._NOT_FOUND_USER));
+    }
+
+    /**
+     * 리뷰 조회
+     * @param reviewId
+     * @return Review
+     */
+    private Review getReviewById(Long reviewId) {
+        return reviewRepository.findById(reviewId)
+                .orElseThrow(() -> new ApiException(ErrorStatus._NOT_FOUND_REVIEW));
+    }
+
+    /**
+     * 사용자 권한 확인
      * @param reviewUserId
      * @param authUserId
      */
-    private void isValidUser(Long reviewUserId, Long authUserId) {
+    private void validateUserPermission(Long reviewUserId, Long authUserId) {
         if (!reviewUserId.equals(authUserId)) {
-            throw new IllegalArgumentException("본인의 리뷰만 수정할 수 있습니다.");
+            throw new ApiException(ErrorStatus._BAD_REQUEST_UPDATE_REVIEW);
         }
     }
-
 }
